# Unroll some of the adler checksum for avx2