Similar to what's done for vmx, avx512, and sse4, let's unroll some of this checksum since it's a commutative checksum. We take advantage of ILP and do more intermediate sums before rolling them back together for the finalization of the checksum.

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- Refactor
  - Optimized Adler-32 on x86 with AVX2, adding a 64-byte processing path and improving 32-byte handling.
  - Delivers faster checksum throughput on large buffers, with additional gains when checksum and copy run together.
  - No changes to user-facing behavior or APIs.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->