@KungFuJesus 2025-08-16
On a meager U class Haswell chip:

## Before
```
-----------------------------------------------------------------
Benchmark                       Time             CPU   Iterations
-----------------------------------------------------------------
adler32/native/1             5.20 ns         5.19 ns    135075246
adler32/native/8             8.37 ns         8.34 ns     84535112
adler32/native/12            10.0 ns         9.99 ns     69944322
adler32/native/16            11.7 ns         11.6 ns     60030516
adler32/native/32            10.9 ns         10.9 ns     64211571
adler32/native/64            12.5 ns         12.5 ns     56268038
adler32/native/512           29.5 ns         29.5 ns     23825567
adler32/native/4096           186 ns          186 ns      3681299
adler32/native/32768         1415 ns         1412 ns       496706
adler32/native/262144       14151 ns        14119 ns        49597
adler32/native/4194304     313660 ns       311133 ns         2210
```

## After
```
-----------------------------------------------------------------
Benchmark                       Time             CPU   Iterations
-----------------------------------------------------------------
adler32/native/1             5.13 ns         5.12 ns    136980626
adler32/native/8             8.29 ns         8.28 ns     84562446
adler32/native/12            9.97 ns         9.95 ns     70399291
adler32/native/16            12.4 ns         12.3 ns     59158817
adler32/native/32            10.7 ns         10.7 ns     65591363
adler32/native/64            12.0 ns         12.0 ns     58241380
adler32/native/512           27.9 ns         27.8 ns     25148383
adler32/native/4096           160 ns          160 ns      4374342
adler32/native/32768         1275 ns         1272 ns       552931
adler32/native/262144       11841 ns        11816 ns        59249
adler32/native/4194304     292282 ns       289905 ns         2333
```

On HEDT class hardware (Cascake Lake X):

## Before
```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
adler32/avx2/1             3.27 ns         3.27 ns    214157823
adler32/avx2/8             5.39 ns         5.39 ns    129875561
adler32/avx2/12            6.73 ns         6.73 ns    103986274
adler32/avx2/16            8.12 ns         8.13 ns     86145285
adler32/avx2/32            7.71 ns         7.71 ns     90711137
adler32/avx2/64            8.21 ns         8.21 ns     85186661
adler32/avx2/512           18.2 ns         18.2 ns     38419988
adler32/avx2/4096           119 ns          119 ns      5905768
adler32/avx2/32768          949 ns          949 ns       736810
adler32/avx2/262144        8403 ns         8406 ns        83277
adler32/avx2/4194304     223228 ns       223337 ns         3136
```

## After
```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
adler32/avx2/1             3.42 ns         3.42 ns    204842982
adler32/avx2/8             6.85 ns         6.85 ns    102142369
adler32/avx2/12            8.81 ns         8.81 ns     79447535
adler32/avx2/16            8.02 ns         8.01 ns     87344982
adler32/avx2/32            7.61 ns         7.62 ns     91836008
adler32/avx2/64            8.05 ns         8.05 ns     86905783
adler32/avx2/512           17.1 ns         17.1 ns     40899431
adler32/avx2/4096           103 ns          103 ns      6810843
adler32/avx2/32768          869 ns          869 ns       804832
adler32/avx2/262144        7638 ns         7640 ns        91662
adler32/avx2/4194304     207382 ns       207414 ns         3435
```